requirements.txt
absl-py==2.1.0
addict==2.4.0
asttokens==3.0.0
attrs==25.1.0
backcall==0.2.0
blinker==1.8.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.1.1
cycler==0.12.1
Cython==0.29.37
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.2.1
egl_probe==1.0.2
executing==2.2.0
fasteners==0.15
fastjsonschema==2.21.1
filelock==3.16.1
Flask==3.0.3
fonttools==4.56.0
free-mujoco-py==2.1.6
fsspec==2025.2.0
glfw==1.12.0
google-auth==2.38.0
google-auth-oauthlib==1.0.0
grpcio==1.70.0
gym==0.26.2
gym-notices==0.0.8
h5py==3.11.0
hidapi==0.14.0.post4
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipython==8.12.3
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
llvmlite==0.36.0
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
monotonic==1.6
mpmath==1.3.0
narwhals==1.28.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
numba==0.53.1
numpy==1.23.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
open3d==0.19.0
packaging==24.2
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
plotly==6.0.0
prompt_toolkit==3.0.50
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybullet-svl==3.1.6.4
pycparser==2.22
pygame==2.6.1
Pygments==2.19.1
pyparsing==3.1.4
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
retrying==1.3.4
rpds-py==0.20.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
six==1.17.0
stack-data==0.6.3
sympy==1.13.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
threadpoolctl==3.5.0
tqdm==4.67.1
traitlets==5.14.3
triton==3.0.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.6
widgetsnbextension==4.0.13
zipp==3.20.2